check description
although not mentioned on the previous push, I added a new mod, "Cyborg Class Palanquin Mech", this mod does that and only that. (base portrait made by Hedera herself, damaged and blink were made by Generic, aka, me).
-new damaged portrait for Palanquin mech

repurposed the mod that added Vanilla's Secret Squad's weapons to the weapon deck, now it's a mod that aims to tweak the whole squad, with one of the features being the original purpose of the mod, other functions include:
-new sprites for the mechs that change their eye color, so it changes with different palettes (can turn on and off from the "mod configuration" menu)

-now the Techno-Beetle is immune to smoke

-the Techno-Scarab will have a new palette that looks more like Vanilla's Alpha Scarab

-now the Techno-Hornet has the big shield sprite when being shielded